Theme color는 따로 분리함
