+++ src/core/GlobalStyle/theme.js
@@ -1,0 +1,20 @@
+const THEME_COLOR = {
+  yellow: {
+    main: yellow100,
+    sub: yellow100,
+  },
+  red: {
+    main: red100,
+    sub: red100,
+  },
+  gray: {
+    main: gray200,
+    sub: gray200,
+  },
+  light: {
+    main: gray450,
+    sub: gray100,
+  },
+};
+
+export { THEME_COLOR };
